fixes after @vctecc review
