+++ senders/alerts/base.py
@@ -1,24 +1,19 @@
 import abc
 
 from senders.celery_app import app
-from senders.services.auth import AuthServiceBase
-from senders.services.ugc import UGCServiceBase
 
 
 class BaseAlert(abc.ABC):
 
-    def __init__(self, template: str, channels: list[str], auth_service: AuthServiceBase, ugc_service: UGCServiceBase):
+    def __init__(self, template: str, channels: list[str]):
         self.template = template
         self.channels = channels
-        self.auth_service = auth_service
-        self.ugc_service = ugc_service
 
     @abc.abstractmethod
     def send(self):
         pass
 
     def _send(self, to: str, data: dict) -> None:
-
         for channel in self.channels:
             payload = self._get_payload(to, channel, data)
             app.send_task(channel, kwargs=payload)
@@ -31,4 +26,3 @@ class BaseAlert(abc.ABC):
             "type": self.template,
             "payload": data
         }
-
